move in-memory storage out of tests
