+++ src/sql_engine/src/catalog_manager/mod.rs
@@ -12,22 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{catalog_manager::metadata::DataDefinition, ColumnDefinition, TableDefinition};
+use crate::{catalog_manager::data_definition::DataDefinition, ColumnDefinition, TableDefinition};
 use kernel::{Object, Operation, SystemError, SystemResult};
+use std::path::PathBuf;
 use std::sync::atomic::{AtomicU64, Ordering};
-use storage::{DatabaseCatalog, ReadCursor, Row, SledDatabaseCatalog, StorageError};
+use storage::{InMemoryDatabaseCatalog, PersistentDatabaseCatalog, ReadCursor, Row, Storage, StorageError};
 
-mod metadata;
+mod data_definition;
 
 pub struct CatalogManager {
     key_id_generator: AtomicU64,
-    persistent: Box<dyn DatabaseCatalog>,
+    data_storage: Box<dyn Storage>,
     data_definition: DataDefinition,
 }
 
-impl CatalogManager {
-    pub fn default() -> SystemResult<Self> {
-        Self::new(Box::new(SledDatabaseCatalog::default()))
+impl Default for CatalogManager {
+    fn default() -> CatalogManager {
+        Self::in_memory().expect("no errors")
     }
 }
 
@@ -37,23 +38,24 @@ unsafe impl Sync for CatalogManager {}
 const DEFAULT_CATALOG: &'_ str = "public";
 
 impl CatalogManager {
-    pub fn new(persistent: Box<dyn DatabaseCatalog>) -> SystemResult<Self> {
-        match persistent.create_namespace_with_objects("system", vec!["columns"]) {
-            Ok(()) => {
-                let definition = DataDefinition::in_memory();
-                definition.create_catalog(DEFAULT_CATALOG);
-                Ok(Self {
-                    key_id_generator: AtomicU64::default(),
-                    persistent,
-                    data_definition: definition,
-                })
-            }
-            Err(StorageError::SystemError(e)) => Err(e),
-            Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
-                Operation::Create,
-                Object::Schema("system"),
-            )),
-        }
+    pub fn in_memory() -> SystemResult<CatalogManager> {
+        let data_definition = DataDefinition::in_memory();
+        data_definition.create_catalog(DEFAULT_CATALOG);
+        Ok(Self {
+            key_id_generator: AtomicU64::default(),
+            data_storage: Box::new(InMemoryDatabaseCatalog::default()),
+            data_definition,
+        })
+    }
+
+    pub fn persistent(path: PathBuf) -> SystemResult<CatalogManager> {
+        let data_definition = DataDefinition::persistent(&path)?;
+        data_definition.create_catalog(DEFAULT_CATALOG);
+        Ok(Self {
+            key_id_generator: AtomicU64::default(),
+            data_storage: Box::new(PersistentDatabaseCatalog::new(path)),
+            data_definition,
+        })
     }
 
     pub fn next_key_id(&self) -> u64 {
@@ -76,7 +78,7 @@ impl CatalogManager {
 
     pub fn create_schema(&self, schema_name: &str) -> SystemResult<()> {
         self.data_definition.create_schema(DEFAULT_CATALOG, schema_name);
-        match self.persistent.create_namespace(schema_name) {
+        match self.data_storage.create_namespace(schema_name) {
             Ok(()) => Ok(()),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -88,7 +90,7 @@ impl CatalogManager {
 
     pub fn drop_schema(&self, schema_name: &str) -> SystemResult<()> {
         self.data_definition.drop_schema(DEFAULT_CATALOG, schema_name);
-        match self.persistent.drop_namespace(schema_name) {
+        match self.data_storage.drop_namespace(schema_name) {
             Ok(()) => Ok(()),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -106,7 +108,7 @@ impl CatalogManager {
     ) -> SystemResult<()> {
         self.data_definition
             .create_table(DEFAULT_CATALOG, schema_name, table_name, column_definitions);
-        match self.persistent.create_tree(schema_name, table_name) {
+        match self.data_storage.create_tree(schema_name, table_name) {
             Ok(()) => Ok(()),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -125,7 +127,7 @@ impl CatalogManager {
     pub fn drop_table(&self, schema_name: &str, table_name: &str) -> SystemResult<()> {
         self.data_definition
             .drop_table(DEFAULT_CATALOG, schema_name, table_name);
-        match self.persistent.drop_tree(schema_name, table_name) {
+        match self.data_storage.drop_tree(schema_name, table_name) {
             Ok(()) => Ok(()),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -137,7 +139,7 @@ impl CatalogManager {
 
     pub fn insert_into(&self, schema_name: &str, table_name: &str, values: Vec<Row>) -> SystemResult<usize> {
         log::debug!("{:#?}", values);
-        match self.persistent.write(schema_name, table_name, values) {
+        match self.data_storage.write(schema_name, table_name, values) {
             Ok(size) => Ok(size),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -148,7 +150,7 @@ impl CatalogManager {
     }
 
     pub fn table_scan(&self, schema_name: &str, table_name: &str) -> SystemResult<ReadCursor> {
-        match self.persistent.read(schema_name, table_name) {
+        match self.data_storage.read(schema_name, table_name) {
             Ok(read) => Ok(read),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -159,7 +161,7 @@ impl CatalogManager {
     }
 
     pub fn update_all(&self, schema_name: &str, table_name: &str, rows: Vec<Row>) -> SystemResult<usize> {
-        match self.persistent.write(schema_name, table_name, rows) {
+        match self.data_storage.write(schema_name, table_name, rows) {
             Ok(size) => Ok(size),
             Err(StorageError::SystemError(error)) => Err(error),
             Err(StorageError::RuntimeCheckError) => Err(SystemError::bug_in_sql_engine(
@@ -170,10 +172,10 @@ impl CatalogManager {
     }
 
     pub fn delete_all_from(&self, schema_name: &str, table_name: &str) -> SystemResult<usize> {
-        match self.persistent.read(schema_name, table_name) {
+        match self.data_storage.read(schema_name, table_name) {
             Ok(reads) => {
                 let keys = reads.map(Result::unwrap).map(|(key, _)| key).collect();
-                match self.persistent.delete(schema_name, table_name, keys) {
+                match self.data_storage.delete(schema_name, table_name, keys) {
                     Ok(len) => Ok(len),
                     _ => unreachable!(
                         "all errors that make code fall in here should have been handled in read operation"
