+++ src/storage/src/lib.rs
@@ -15,20 +15,25 @@
 extern crate kernel;
 extern crate log;
 
-use kernel::SystemError;
+use kernel::{SystemError, SystemResult};
 use representation::Binary;
-use sled::{Db as NameSpace, Error as SledError};
-use std::{
-    collections::HashMap,
-    path::PathBuf,
-    sync::{Arc, RwLock},
-};
 
 pub type Row = (Key, Values);
 pub type Key = Binary;
 pub type Values = Binary;
-pub type ReadCursor = Box<dyn Iterator<Item = Result<Row, SystemError>>>;
+pub type ReadCursor = Box<dyn Iterator<Item = SystemResult<Row>>>;
 pub type StorageResult<T> = std::result::Result<T, StorageError>;
+
+mod in_memory;
+mod persistent;
+
+pub use crate::in_memory::InMemoryDatabaseCatalog;
+pub use crate::persistent::PersistentDatabaseCatalog;
+
+pub enum InitStatus {
+    Created,
+    Loaded,
+}
 
 #[derive(Debug, PartialEq)]
 pub enum StorageError {
@@ -36,570 +41,21 @@ pub enum StorageError {
     SystemError(SystemError),
 }
 
-pub trait DatabaseCatalog {
-    fn create_namespace_with_objects(&self, namespace: &str, object_names: Vec<&str>) -> StorageResult<()>;
-
+pub trait Storage {
     fn create_namespace(&self, namespace: &str) -> StorageResult<()>;
 
     fn drop_namespace(&self, namespace: &str) -> StorageResult<()>;
 
-    fn is_namespace_exists(&self, namespace: &str) -> bool;
-
     fn create_tree(&self, namespace: &str, object_name: &str) -> StorageResult<()>;
 
     fn drop_tree(&self, namespace: &str, object_name: &str) -> StorageResult<()>;
-
-    fn is_tree_exists(&self, namespace: &str, object_name: &str) -> bool;
 
     fn write(&self, namespace: &str, object_name: &str, values: Vec<Row>) -> StorageResult<usize>;
 
     fn read(&self, namespace: &str, object_name: &str) -> StorageResult<ReadCursor>;
 
     fn delete(&self, namespace: &str, object_name: &str, keys: Vec<Key>) -> StorageResult<usize>;
-
-    fn check_for_object(&self, namespace: &str, object_name: &str) -> StorageResult<()>;
-}
-
-pub struct SledErrorMapper;
-
-impl SledErrorMapper {
-    fn map(error: SledError) -> SystemError {
-        match error {
-            SledError::CollectionNotFound(system_file) => SystemError::unrecoverable(format!(
-                "System file [{}] can't be found",
-                String::from_utf8(system_file.to_vec()).expect("name of system file")
-            )),
-            SledError::Unsupported(operation) => {
-                SystemError::unrecoverable(format!("Unsupported operation [{}] was used on Sled", operation))
-            }
-            SledError::Corruption { at, bt: _bt } => {
-                if let Some(at) = at {
-                    SystemError::unrecoverable(format!("Sled encountered corruption at {}", at))
-                } else {
-                    SystemError::unrecoverable("Sled encountered corruption".to_owned())
-                }
-            }
-            SledError::ReportableBug(description) => {
-                SystemError::unrecoverable(format!("Sled encountered reportable BUG: {}", description))
-            }
-            SledError::Io(error) => SystemError::io(error),
-        }
-    }
-}
-
-pub enum InitStatus {
-    Created,
-    Loaded,
-}
-
-#[derive(Default)]
-pub struct SledDatabaseCatalog {
-    path: Option<PathBuf>,
-    namespaces: RwLock<HashMap<String, Arc<NameSpace>>>,
-}
-
-impl SledDatabaseCatalog {
-    pub fn new(path: PathBuf) -> SledDatabaseCatalog {
-        SledDatabaseCatalog {
-            path: Some(path),
-            namespaces: RwLock::new(HashMap::new()),
-        }
-    }
-
-    pub fn init(&self, namespace_name: &str) -> StorageResult<InitStatus> {
-        eprintln!("{:?}", self.path);
-        if let Some(path) = self.path.as_ref() {
-            let path_to_namespace = PathBuf::from(path).join(namespace_name);
-            match sled::open(path_to_namespace) {
-                Ok(namespace) => {
-                    let recovered = namespace.was_recovered();
-                    self.namespaces
-                        .write()
-                        .expect("to acquire write lock")
-                        .insert(namespace_name.to_owned(), Arc::new(namespace));
-                    if recovered {
-                        Ok(InitStatus::Loaded)
-                    } else {
-                        Ok(InitStatus::Created)
-                    }
-                }
-                Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-            }
-        } else {
-            Err(StorageError::RuntimeCheckError)
-        }
-    }
-
-    fn new_namespace(&self, namespace: &str) -> StorageResult<Arc<NameSpace>> {
-        if self
-            .namespaces
-            .read()
-            .expect("to acquire read lock")
-            .contains_key(namespace)
-        {
-            Err(StorageError::RuntimeCheckError)
-        } else {
-            match sled::Config::default().temporary(true).open() {
-                Ok(database) => {
-                    let database = Arc::new(database);
-                    self.namespaces
-                        .write()
-                        .expect("to acquire write lock")
-                        .insert(namespace.to_owned(), database.clone());
-                    Ok(database)
-                }
-                Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-            }
-        }
-    }
-}
-
-impl DatabaseCatalog for SledDatabaseCatalog {
-    fn create_namespace_with_objects(&self, namespace: &str, object_names: Vec<&str>) -> StorageResult<()> {
-        let namespace = self.new_namespace(namespace)?;
-        for object_name in object_names {
-            match namespace.open_tree(object_name) {
-                Ok(_object) => (),
-                Err(error) => return Err(StorageError::SystemError(SledErrorMapper::map(error))),
-            }
-        }
-        Ok(())
-    }
-
-    fn create_namespace(&self, namespace: &str) -> StorageResult<()> {
-        self.new_namespace(namespace).map(|_| ())
-    }
-
-    fn drop_namespace(&self, namespace: &str) -> StorageResult<()> {
-        match self
-            .namespaces
-            .write()
-            .expect("to acquire write lock")
-            .remove(namespace)
-        {
-            Some(namespace) => {
-                drop(namespace);
-                Ok(())
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn create_tree(&self, namespace: &str, object_name: &str) -> StorageResult<()> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => {
-                if namespace.tree_names().contains(&(object_name.into())) {
-                    Err(StorageError::RuntimeCheckError)
-                } else {
-                    match namespace.open_tree(object_name) {
-                        Ok(_object) => Ok(()),
-                        Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                    }
-                }
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn drop_tree(&self, namespace: &str, object_name: &str) -> StorageResult<()> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => match namespace.drop_tree(object_name.as_bytes()) {
-                Ok(true) => Ok(()),
-                Ok(false) => Err(StorageError::RuntimeCheckError),
-                Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-            },
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn write(&self, namespace: &str, object_name: &str, rows: Vec<Row>) -> StorageResult<usize> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => {
-                if namespace.tree_names().contains(&(object_name.into())) {
-                    match namespace.open_tree(object_name) {
-                        Ok(object) => {
-                            let mut written_rows = 0;
-                            for (key, values) in rows {
-                                match object
-                                    .insert::<sled::IVec, sled::IVec>(key.to_bytes().into(), values.to_bytes().into())
-                                {
-                                    Ok(_) => written_rows += 1,
-                                    Err(error) => return Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                                }
-                            }
-                            Ok(written_rows)
-                        }
-                        Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                    }
-                } else {
-                    Err(StorageError::RuntimeCheckError)
-                }
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn read(&self, namespace: &str, object_name: &str) -> StorageResult<ReadCursor> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => {
-                if namespace.tree_names().contains(&(object_name.into())) {
-                    match namespace.open_tree(object_name) {
-                        Ok(object) => Ok(Box::new(object.iter().map(|item| match item {
-                            Ok((key, values)) => {
-                                Ok((Binary::with_data(key.to_vec()), Binary::with_data(values.to_vec())))
-                            }
-                            Err(error) => Err(SledErrorMapper::map(error)),
-                        }))),
-                        Err(error) => Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                    }
-                } else {
-                    Err(StorageError::RuntimeCheckError)
-                }
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn delete(&self, namespace: &str, object_name: &str, keys: Vec<Key>) -> StorageResult<usize> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => {
-                if namespace.tree_names().contains(&(object_name.into())) {
-                    let mut deleted = 0;
-                    match namespace.open_tree(object_name) {
-                        Ok(object) => {
-                            for key in keys {
-                                match object.remove(key.to_bytes()) {
-                                    Ok(_) => deleted += 1,
-                                    Err(error) => return Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                                }
-                            }
-                        }
-                        Err(error) => return Err(StorageError::SystemError(SledErrorMapper::map(error))),
-                    }
-                    Ok(deleted)
-                } else {
-                    Err(StorageError::RuntimeCheckError)
-                }
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
-
-    fn is_namespace_exists(&self, namespace: &str) -> bool {
-        self.namespaces
-            .read()
-            .expect("to acquire read lock")
-            .contains_key(namespace)
-    }
-
-    fn is_tree_exists(&self, namespace: &str, object_name: &str) -> bool {
-        self.check_for_object(namespace, object_name).is_ok()
-    }
-
-    fn check_for_object(&self, namespace: &str, object_name: &str) -> StorageResult<()> {
-        match self.namespaces.read().expect("to acquire read lock").get(namespace) {
-            Some(namespace) => {
-                if namespace.tree_names().contains(&(object_name.into())) {
-                    Ok(())
-                } else {
-                    Err(StorageError::RuntimeCheckError)
-                }
-            }
-            None => Err(StorageError::RuntimeCheckError),
-        }
-    }
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use kernel::SystemResult;
-
-    type Storage = SledDatabaseCatalog;
-
-    #[rstest::fixture]
-    fn storage() -> Storage {
-        Storage::default()
-    }
-
-    #[rstest::fixture]
-    fn with_namespace(storage: Storage) -> Storage {
-        storage.create_namespace("namespace").expect("namespace created");
-        storage
-    }
-
-    #[rstest::fixture]
-    fn with_object(with_namespace: Storage) -> Storage {
-        with_namespace
-            .create_tree("namespace", "object_name")
-            .expect("object created");
-        with_namespace
-    }
-
-    #[cfg(test)]
-    mod sled_error_mapper {
-        use super::*;
-        use sled::DiskPtr;
-        use std::io::{Error, ErrorKind};
-
-        #[test]
-        fn collection_not_found() {
-            assert_eq!(
-                SledErrorMapper::map(SledError::CollectionNotFound(sled::IVec::from("test"))),
-                SystemError::unrecoverable("System file [test] can't be found".to_owned())
-            )
-        }
-
-        #[test]
-        fn unsupported() {
-            assert_eq!(
-                SledErrorMapper::map(SledError::Unsupported("NOT_SUPPORTED".to_owned())),
-                SystemError::unrecoverable("Unsupported operation [NOT_SUPPORTED] was used on Sled".to_owned())
-            )
-        }
-
-        #[test]
-        fn corruption_with_position() {
-            let at = DiskPtr::Inline(900);
-            assert_eq!(
-                SledErrorMapper::map(SledError::Corruption { at: Some(at), bt: () }),
-                SystemError::unrecoverable(format!("Sled encountered corruption at {}", at))
-            )
-        }
-
-        #[test]
-        fn corruption_without_position() {
-            assert_eq!(
-                SledErrorMapper::map(SledError::Corruption { at: None, bt: () }),
-                SystemError::unrecoverable("Sled encountered corruption".to_owned())
-            )
-        }
-
-        #[test]
-        fn reportable_bug() {
-            let description = "SOME_BUG_HERE";
-            assert_eq!(
-                SledErrorMapper::map(SledError::ReportableBug(description.to_owned())),
-                SystemError::unrecoverable(format!("Sled encountered reportable BUG: {}", description))
-            );
-        }
-
-        #[test]
-        fn io() {
-            assert_eq!(
-                SledErrorMapper::map(SledError::Io(Error::new(ErrorKind::Other, "oh no!"))),
-                SystemError::io(Error::new(ErrorKind::Other, "oh no!"))
-            )
-        }
-    }
-
-    #[cfg(test)]
-    mod namespace {
-        use super::*;
-
-        #[rstest::rstest]
-        fn create_namespace_with_objects(storage: Storage) {
-            assert_eq!(
-                storage.create_namespace_with_objects("namespace", vec!["object_1", "object_2"]),
-                Ok(())
-            );
-
-            assert!(storage.is_tree_exists("namespace", "object_1"));
-            assert!(storage.is_tree_exists("namespace", "object_2"));
-        }
-
-        #[rstest::rstest]
-        fn create_namespaces_with_different_names(storage: Storage) {
-            assert_eq!(storage.create_namespace("namespace_1"), Ok(()));
-            assert_eq!(storage.create_namespace("namespace_2"), Ok(()));
-        }
-
-        #[rstest::rstest]
-        fn drop_namespace(with_namespace: Storage) {
-            assert_eq!(with_namespace.drop_namespace("namespace"), Ok(()));
-            assert_eq!(with_namespace.create_namespace("namespace"), Ok(()));
-        }
-
-        #[rstest::rstest]
-        fn dropping_namespace_drops_objects_in_it(with_namespace: Storage) {
-            with_namespace
-                .create_tree("namespace", "object_name_1")
-                .expect("object created");
-            with_namespace
-                .create_tree("namespace", "object_name_2")
-                .expect("object created");
-
-            assert_eq!(with_namespace.drop_namespace("namespace"), Ok(()));
-            assert_eq!(with_namespace.create_namespace("namespace"), Ok(()));
-            assert_eq!(with_namespace.create_tree("namespace", "object_name_1"), Ok(()));
-            assert_eq!(with_namespace.create_tree("namespace", "object_name_2"), Ok(()));
-        }
-    }
-
-    #[cfg(test)]
-    mod create_object {
-        use super::*;
-
-        #[rstest::rstest]
-        fn create_objects_with_different_names(with_namespace: Storage) {
-            assert_eq!(with_namespace.create_tree("namespace", "object_name_1"), Ok(()));
-            assert_eq!(with_namespace.create_tree("namespace", "object_name_2"), Ok(()));
-        }
-
-        #[rstest::rstest]
-        fn create_object_with_the_same_name_in_different_namespaces(storage: Storage) {
-            storage.create_namespace("namespace_1").expect("namespace created");
-            storage.create_namespace("namespace_2").expect("namespace created");
-            assert_eq!(storage.create_tree("namespace_1", "object_name"), Ok(()));
-            assert_eq!(storage.create_tree("namespace_2", "object_name"), Ok(()));
-        }
-    }
-
-    #[cfg(test)]
-    mod drop_object {
-        use super::*;
-
-        #[rstest::rstest]
-        fn drop_object(with_object: Storage) {
-            assert_eq!(with_object.drop_tree("namespace", "object_name"), Ok(()));
-            assert_eq!(with_object.create_tree("namespace", "object_name"), Ok(()));
-        }
-    }
-
-    #[cfg(test)]
-    mod operations_on_object {
-        use super::*;
-
-        #[rstest::rstest]
-        fn insert_row_into_object(with_object: Storage) {
-            assert_eq!(
-                with_object.write("namespace", "object_name", as_rows(vec![(1u8, vec!["123"])])),
-                Ok(1)
-            );
-
-            assert_eq!(
-                with_object
-                    .read("namespace", "object_name")
-                    .map(|iter| iter.collect::<Vec<SystemResult<Row>>>()),
-                Ok(as_read_cursor(vec![(1u8, vec!["123"])]).collect())
-            );
-        }
-
-        #[rstest::rstest]
-        fn insert_many_rows_into_object(with_object: Storage) {
-            with_object
-                .write("namespace", "object_name", as_rows(vec![(1u8, vec!["123"])]))
-                .expect("values are written");
-            with_object
-                .write("namespace", "object_name", as_rows(vec![(2u8, vec!["456"])]))
-                .expect("values are written");
-
-            assert_eq!(
-                with_object
-                    .read("namespace", "object_name")
-                    .map(|iter| iter.collect::<Vec<SystemResult<Row>>>()),
-                Ok(as_read_cursor(vec![(1u8, vec!["123"]), (2u8, vec!["456"])]).collect())
-            );
-        }
-
-        #[rstest::rstest]
-        fn delete_some_records_from_object(with_object: Storage) {
-            with_object
-                .write(
-                    "namespace",
-                    "object_name",
-                    as_rows(vec![(1u8, vec!["123"]), (2u8, vec!["456"]), (3u8, vec!["789"])]),
-                )
-                .expect("write occurred");
-
-            assert_eq!(
-                with_object.delete("namespace", "object_name", as_keys(vec![2u8])),
-                Ok(1)
-            );
-
-            assert_eq!(
-                with_object
-                    .read("namespace", "object_name")
-                    .map(|iter| iter.collect::<Vec<SystemResult<Row>>>()),
-                Ok(as_read_cursor(vec![(1u8, vec!["123"]), (3u8, vec!["789"])]).collect())
-            );
-        }
-
-        #[rstest::rstest]
-        fn select_all_from_object_with_many_columns(with_object: Storage) {
-            with_object
-                .write("namespace", "object_name", as_rows(vec![(1u8, vec!["1", "2", "3"])]))
-                .expect("write occurred");
-
-            assert_eq!(
-                with_object
-                    .read("namespace", "object_name")
-                    .map(|iter| iter.collect::<Vec<SystemResult<Row>>>()),
-                Ok(as_read_cursor(vec![(1u8, vec!["1", "2", "3"])]).collect())
-            );
-        }
-
-        #[rstest::rstest]
-        fn insert_multiple_rows(with_object: Storage) {
-            with_object
-                .write(
-                    "namespace",
-                    "object_name",
-                    as_rows(vec![
-                        (1u8, vec!["1", "2", "3"]),
-                        (2u8, vec!["4", "5", "6"]),
-                        (3u8, vec!["7", "8", "9"]),
-                    ]),
-                )
-                .expect("write occurred");
-
-            assert_eq!(
-                with_object
-                    .read("namespace", "object_name")
-                    .map(|iter| iter.collect::<Vec<SystemResult<Row>>>()),
-                Ok(as_read_cursor(vec![
-                    (1u8, vec!["1", "2", "3"]),
-                    (2u8, vec!["4", "5", "6"]),
-                    (3u8, vec!["7", "8", "9"])
-                ])
-                .collect()),
-            );
-        }
-    }
-
-    fn as_rows(items: Vec<(u8, Vec<&'static str>)>) -> Vec<Row> {
-        items
-            .into_iter()
-            .map(|(key, values)| {
-                let k = Binary::with_data(key.to_be_bytes().to_vec());
-                let v = Binary::with_data(
-                    values
-                        .into_iter()
-                        .map(|s| s.as_bytes())
-                        .collect::<Vec<&[u8]>>()
-                        .join(&b'|'),
-                );
-                (k, v)
-            })
-            .collect()
-    }
-
-    fn as_keys(items: Vec<u8>) -> Vec<Key> {
-        items
-            .into_iter()
-            .map(|key| Binary::with_data(key.to_be_bytes().to_vec()))
-            .collect()
-    }
-
-    fn as_read_cursor(items: Vec<(u8, Vec<&'static str>)>) -> ReadCursor {
-        Box::new(items.into_iter().map(|(key, values)| {
-            let k = key.to_be_bytes().to_vec();
-            let v = values
-                .into_iter()
-                .map(|s| s.as_bytes())
-                .collect::<Vec<&[u8]>>()
-                .join(&b'|');
-            Ok((Binary::with_data(k), Binary::with_data(v)))
-        }))
-    }
-}
+mod tests;
